Eliminate scope.h
Map and for statements no longer store their current value in the scope's internal data structures. Instead, they store an AstVariable node in the scope and update that with the current value at evaluation-time. Other nodes access the current value by storing a pointer to the AstVariable. As a result, parts of the scope no longer appear in the AST, and the scope is now only active at parse-time.

The downside is that the debug code no longer has access to all variable names.